+++ extension/src/content-script/wybin/content.ts
@@ -3,8 +3,30 @@ import { addFlagsHome } from "./pages/home";
 import { addFlagsParticipants } from "./pages/participants";
 import { addFlagsPlayers } from "./pages/players";
 import { addFlagsSchedule } from "./pages/schedule";
-import { addFlagsTeams } from "./pages/teams";
+import { addFlagsTeams, updateFlagTeams } from "./pages/teams";
 import { addFlagsTournamentManagement } from "./pages/tournamentManagement";
+
+const bodyObserver = new MutationObserver((mutations) => {
+    const overlay = document.querySelector("body > div.cdk-overlay-container");
+    if (overlay) {
+        overlayObserver.observe(overlay, { childList: true });
+
+        const teamEl = (mutations[0].addedNodes?.[0] as HTMLElement)?.querySelector(".team");
+        if (teamEl) {
+            updateFlagTeams(teamEl as HTMLElement);
+        }
+    }
+});
+
+const overlayObserver = new MutationObserver((mutation) => {
+    if (mutation[0].target) {
+        execOverlay(mutation[0].target as HTMLElement);
+    }
+});
+
+const execOverlay = (parent: HTMLElement) => {
+    updateFlagTeams(parent);
+};
 
 const contentObserver = new MutationObserver(() => {
     exec();
@@ -20,7 +42,7 @@ let content: HTMLElement | undefined = undefined;
 
 export const exec = async () => {
     titleObserver.observe(document.querySelector("head > title")!, { childList: true });
-
+    bodyObserver.observe(document.body, { childList: true });
     const pathname = location.pathname;
     if (pathname === "/draft") {
         draftPageEnhance();
@@ -51,7 +73,7 @@ const draftPageEnhance = () => {
     if (!content) {
         return;
     }
-    contentObserver.observe( content, { childList: true });
+    contentObserver.observe(content, { childList: true });
     enhanceDraftPage();
 };